+++ packages/core/src/messagingProtocols/proteus/ProteusService/ProteusService.ts
@@ -1,0 +1,129 @@
+/*
+ * Wire
+ * Copyright (C) 2022 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+import {APIClient} from '@wireapp/api-client/lib/APIClient';
+import {Conversation, NewConversation} from '@wireapp/api-client/lib/conversation';
+import logdown from 'logdown';
+
+import {
+  AddUsersToProteusConversationParams,
+  CreateProteusConversationParams,
+  ProteusServiceConfig,
+  SendProteusMessageParams,
+} from './ProteusService.types';
+import {isClearFromMismatch} from './Utility';
+import {getGenericMessageParams} from './Utility/getGenericMessageParams';
+
+import {PayloadBundleState, SendResult} from '../../../conversation';
+import {MessageService} from '../../../conversation/message/MessageService';
+import {CryptographyService} from '../../../cryptography';
+import {EventHandlerResult} from '../../common.types';
+import {EventHandlerParams, handleBackendEvent} from '../EventHandler';
+
+export class ProteusService {
+  private readonly messageService: MessageService;
+  private readonly logger = logdown('@wireapp/core/ProteusService');
+
+  constructor(
+    private readonly apiClient: APIClient,
+    private readonly cryptographyService: CryptographyService,
+    private readonly config: ProteusServiceConfig,
+  ) {
+    this.messageService = new MessageService(this.apiClient, this.cryptographyService);
+  }
+
+  public async handleEvent(params: Omit<EventHandlerParams, 'cryptographyService'>): EventHandlerResult {
+    return handleBackendEvent({...params, cryptographyService: this.cryptographyService});
+  }
+
+  public async createConversation({
+    conversationData,
+    otherUserIds,
+  }: CreateProteusConversationParams): Promise<Conversation> {
+    let payload: NewConversation;
+    if (typeof conversationData === 'string') {
+      const ids = typeof otherUserIds === 'string' ? [otherUserIds] : otherUserIds;
+
+      payload = {
+        name: conversationData,
+        receipt_mode: null,
+        users: ids ?? [],
+      };
+    } else {
+      payload = conversationData;
+    }
+
+    return this.apiClient.api.conversation.postConversation(payload);
+  }
+
+  public async addUsersToConversation({conversationId, qualifiedUserIds}: AddUsersToProteusConversationParams) {
+    return this.apiClient.api.conversation.postMembers(conversationId, qualifiedUserIds);
+  }
+
+  public async sendMessage({
+    userIds,
+    sendAsProtobuf,
+    conversationId,
+    nativePush,
+    targetMode,
+    payload,
+    onClientMismatch,
+  }: SendProteusMessageParams): Promise<SendResult> {
+    const messageParams = await getGenericMessageParams({
+      apiClient: this.apiClient,
+      sendingClientId: this.apiClient.validatedClientId,
+      conversationId,
+      genericMessage: payload,
+      useQualifiedIds: this.config.useQualifiedIds,
+      options: {
+        userIds,
+        sendAsProtobuf,
+        nativePush,
+        targetMode,
+        onClientMismatch,
+      },
+    });
+
+    const {federated, sendingClientId, recipients, plainText, options} = messageParams;
+    const response = federated
+      ? await this.messageService.sendFederatedMessage(sendingClientId, recipients, plainText, {
+          ...options,
+          onClientMismatch: mismatch => onClientMismatch?.(mismatch, false),
+        })
+      : await this.messageService.sendMessage(sendingClientId, recipients, plainText, {
+          ...options,
+          sendAsProtobuf,
+          onClientMismatch: mismatch => onClientMismatch?.(mismatch, false),
+        });
+
+    if (!response.errored) {
+      if (!isClearFromMismatch(response)) {
+        // We warn the consumer that there is a mismatch that did not prevent message sending
+        await onClientMismatch?.(response, true);
+      }
+      this.logger.log(`Successfully sent Proteus message to conversation '${conversationId.id}'`);
+    }
+
+    return {
+      id: payload.messageId,
+      sentAt: response.time,
+      state: response.errored ? PayloadBundleState.CANCELLED : PayloadBundleState.OUTGOING_SENT,
+    };
+  }
+}